Add node allocation support
This should really be split into two commits, as it provides both
node allocation and moves to using an adapter pool instead of a
subnet pool, but adding node allocation support exposed some
strange behaviour with the nat daemon that necessitated the move
away from a subnet pool. The fusion nat daemon won't nat incoming
requests to vms that aren't on its subnet. It will still nat
outgoing requests, however. We mitigate this by creating and
managing a pool of adapters, each with its own subnet.

Test plan:
- nalloc-init and nalloc-ssh

+++ lib/nalloc/driver/fusion.rb
@@ -1,21 +1,201 @@
 require 'erb'
+require 'json'
+require 'tempfile'
 
 require Nalloc.libpath('nalloc/fusion_support/vmdk_parser')
+require Nalloc.libpath('nalloc/fusion_support/adapter_pool')
+require Nalloc.libpath('nalloc/node')
 
 class Nalloc::Driver::Fusion < Nalloc::Driver
-  CONFIG_PATH      = File.expand_path("~/.nalloc/fusion")
-  SUBNET_POOL_PATH = File.join(CONFIG_PATH, "subnet_pool")
-  VM_STORE_PATH    = File.join(CONFIG_PATH, "vms")
+  CONFIG_DIR        = File.expand_path("~/.nalloc/fusion")
+  DEFAULTS_PATH     = File.join(CONFIG_DIR, "defaults.json")
+  ADAPTER_POOL_PATH = File.join(CONFIG_DIR, "adapter_pool")
+  VM_STORE_PATH     = File.join(CONFIG_DIR, "vms")
+
+  TEMPLATE_DIR              = File.expand_path("../../../templates/fusion",
+                                               File.dirname(__FILE__))
+  VMX_TEMPLATE_PATH         = File.join(TEMPLATE_DIR, "zygote.vmx.erb")
+  IFACES_TEMPLATE_PATH      = File.join(TEMPLATE_DIR, "interfaces.erb")
+  RESOLV_CONF_TEMPLATE_PATH = File.join(TEMPLATE_DIR, "resolv.conf.erb")
+
+  def self.write_defaults(defaults)
+    File.open(DEFAULTS_PATH, 'w+') do |f|
+      f.write(defaults.to_json)
+    end
+
+    nil
+  end
+
+  # Returns default options used during node allocation
+  #
+  # @return [Hash]  'vmdk_path'         => Path to base vmdk
+  #                 'vmx_template_path' => Path to vmx erb template
+  def self.defaults
+    unless @defaults
+      raw_defaults = File.read(DEFAULTS_PATH)
+      @defaults = JSON.parse(raw_defaults)
+      @defaults.freeze
+    end
+
+    @defaults
+  end
+
+  def self.default_on(specs, prop)
+    specs[prop] || self.defaults[prop.to_s]
+  end
+
+  def initialize(opts={})
+    @adapter_pool = opts[:adapter_pool] || \
+                    Nalloc::FusionSupport::AdapterPool.new(ADAPTER_POOL_PATH)
+  end
 
   def name
     "fusion"
   end
 
-  def destroy_cluster
+  # Begins allocation of the cluster specified by *specs*
+  #
+  # Side effects:
+  # - Reserves an adapter from the pool for the lifetime of the cluster.
+  #
+  # @param  [String]  cluster_id  Globally unique cluster id.
+  # @param  [Hash]    specs       Node properties. See #start_allocating_node
+  #
+  # @return [Hash]                Node name => block. See #start_allocating_node
+  def start_allocating_nodes(cluster_id, specs)
+    # .1 is the host adapter, .2 is the nat daemon
+    raise "Subnets are limited to 251 nodes" if specs.length > 251
+
+    return {} if specs.empty?
+
+    adapter = @adapter_pool.acquire
+    raise "Failed to allocate an adapter" unless adapter
+
+    nodes    = {}
+    network  = adapter["subnet"].split(".")[0, 3].join(".")
+    gateway  = network + ".2"
+    bcast    = network + ".255"
+    host_ctr = 3
+    specs.each do |name, spec|
+      nodes[name] =
+        start_allocating_node(cluster_id, name, spec,
+                              :ipaddr     => network + ".#{host_ctr}",
+                              :netmask    => adapter["netmask"],
+                              :gateway    => gateway,
+                              :broadcast  => bcast,
+                              :dns_server => gateway,
+                              :adapter    => adapter["adapter_id"])
+      host_ctr +=1
+    end
+
+    nodes
+  rescue
+    adapter_pool.release(adapter) if adapter
+
+    raise
+  end
+
+  # Starts allocating a node that belongs to +cluster_id+ and is specified by
+  # +specs+ and +props+
+  #
+  # Side effects:
+  # - Creates a directory to house the vm if it does not already exist. This
+  #   directory lives at VM_STORE_PATH/*cluster_id*
+  #
+  # @param  [String]  cluster_id  Globally unique cluster id.
+  # @param  [String]  name        Node name.
+  # @param  [Hash]    specs       Map of caller supplied node properties.
+  #                                 :vmdk_path  => Optional.
+  #                                 :vmx_path   => Optional.
+  #                                 :memsize_MB => Optional. RAM to give the vm.
+  #                                 :root_pass => Required. Root pass is
+  #                                   needed to configure networking, etc.
+  #                                 :username  => Required. Initial user.
+  #                                 :ssh_key_name => Required. Private ssh key
+  #                                   name. Corresponding public key will be
+  #                                   added to initial user's authorized_keys
+  #                                   file.
+  #
+  # @return block    Finishes allocating node, blocks until completion.
+  def start_allocating_node(cluster_id, name, specs, props)
+    vmdk_path = self.class.default_on(specs, :vmdk_path)
+    raise "No vmdk_path specified for node #{name}" unless vmdk_path
+
+    root_pass = self.class.default_on(specs, :root_pass)
+    raise "No root_pass specified for node #{name}" unless root_pass
+
+    user = specs[:username]
+    raise "No user given for node #{name}" unless specs[:username]
+
+    ssh_key_name = specs[:ssh_key_name]
+    raise "No ssh_key_name given for node #{name}" unless specs[:ssh_key_name]
+
+    private_key_path = Nalloc::Node.find_ssh_key(ssh_key_name)
+    public_key_path  = Nalloc::Node.find_ssh_key("#{ssh_key_name}.pub")
+
+    vmx_template_path = self.class.default_on(specs, :vmx_template_path)
+    vmx_template_path ||= VMX_TEMPLATE_PATH
+
+    cluster_dir = File.join(VM_STORE_PATH, cluster_id)
+    unless File.exist?(cluster_dir)
+      FileUtils.mkdir(cluster_dir)
+    end
+
+    vmx_path = create_vm(vmx_template_path, cluster_dir,
+                         :name          => name,
+                         :vmdk_path     => vmdk_path,
+                         :vmnet_adapter => "vmnet#{props[:adapter]}",
+                         :memsize_MB    => specs[:memsize_MB])
+
+    vmrun("start", vmx_path, "nogui")
+
+    lambda do
+      # Set up network interface
+      ifaces = Tempfile.new("nalloc_ifaces")
+      write_template(IFACES_TEMPLATE_PATH, ifaces.path, props)
+      copy_to_guest(root_pass, vmx_path, ifaces.path, "/etc/network/interfaces",
+                    :mode => "0755")
+
+      # Set up DNS
+      resolv_conf = Tempfile.new("nalloc_resolv_conf")
+      write_template(RESOLV_CONF_TEMPLATE_PATH, resolv_conf.path, props)
+      copy_to_guest(root_pass, vmx_path, resolv_conf.path, "/etc/resolv.conf",
+                    :mode => "0755")
+
+      # Make sure .ssh exists and has correct permissions
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/bin/mkdir", "-p", "/home/#{user}/.ssh")
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/bin/chown", "#{user}:#{user}", "/home/#{user}/.ssh")
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/bin/chmod", "0700", "/home/#{user}/.ssh")
+
+      # Set authorized keys for user
+      guest_auth_keys_path = "/home/#{user}/.ssh/authorized_keys"
+      copy_to_guest(root_pass, vmx_path, public_key_path, guest_auth_keys_path,
+                    :owner => user,
+                    :mode  => "0600")
+
+      # Bring up networking
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/etc/init.d/networking", "restart")
+
+      { "identity"          => vmx_path,
+        "public_ip_address" => props[:ipaddr],
+        "ssh" => {
+          "user" => user,
+          "private_key_name" => ssh_key_name,
+          "public_host_key"  => Nalloc::Node.ssh_public_host_key(props[:ipaddr])
+        }
+      }
+    end
+  end
+
+  def destroy_cluster(cluster_id)
     nil
   end
 
-  # Creates the vm specifed using the supplied vmx template and vmdk.
+  # Creates the vm specifed by *vm_props*.
   #
   # NB: Fusion doesn't explicitly support cloning of vms. We approximate
   # the process using a method similar to the one detailed at
@@ -29,10 +209,10 @@ class Nalloc::Driver::Fusion < Nalloc::Driver
   # @option vm_props  [String]  :name           Required. The vm name.
   # @option vm_props  [String]  :vmdk_path      Required. Path to base vmdk.
   # @option vm_props  [String]  :vmnet_adapter  Required. vmnet0 for example.
-  # @option vm_props  [Integer] :memsize_MB     Optional. Mem size in MB.
+  # @option vm_props  [Integer] :memsize_MB     Optional. Ram allocated to vm.
   # @option vm_props  [String]  :guest_os       Optional. VMware specific guest
   #                                             os tag.
-  # @return nil
+  # @return [String]                     Path to newly created vmx file.
   def create_vm(vmx_template_path, vm_store_path, vm_props)
     vm_dir = File.join(vm_store_path, "#{vm_props[:name]}.vmwarevm")
 
@@ -54,22 +234,20 @@ class Nalloc::Driver::Fusion < Nalloc::Driver
 
     # Write out the vmx file for the vm
     dst_vmx_path = File.join(vm_dir, "#{vm_props[:name]}.vmx")
-    write_vmx_template(vmx_template_path, dst_vmx_path, vm_props)
+    write_template(vmx_template_path, dst_vmx_path, vm_props)
 
     # Finally snapshot the cloned vm. This will force any writes to be
     # performed against the local linked disk instead of the zygote's disk.
     vmrun("snapshot", dst_vmx_path, "base")
 
-    nil
+    dst_vmx_path
   rescue
-    if created_vm_dir
-      FileUtils.rm_rf(vm_dir)
-    end
+    FileUtils.rm_rf(vm_dir) if created_vm_dir
 
     raise
   end
 
-  def write_vmx_template(template_path, dst_path, props)
+  def write_template(template_path, dst_path, props)
     raw_template = File.read(template_path)
     template = ERB.new(raw_template)
     File.open(dst_path, 'w+') do |f|
@@ -86,5 +264,22 @@ class Nalloc::Driver::Fusion < Nalloc::Driver
     raise "vmrun operation failed: #{command.inspect}" unless $?.success?
 
     result
+  end
+
+  def copy_to_guest(root_pass, vmx_path, src_path, dst_path, opts={})
+    vmrun("-gu", "root", "-gp", root_pass, "CopyFileFromHostToGuest",
+          vmx_path, src_path, dst_path)
+
+    if owner = opts[:owner]
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/bin/chown", "#{owner}:#{owner}", dst_path)
+    end
+
+    if mode = opts[:mode]
+      vmrun("-gu", "root", "-gp", root_pass, "runProgramInGuest", vmx_path,
+            "/bin/chmod", mode, dst_path)
+    end
+
+    nil
   end
 end
